+++ mywords/trunk/post_data.php
@@ -44,9 +44,11 @@ while ($row = $db->fetchArray($result)){
     }
     
     // Redes Sociales
-    $bms = array();
-    foreach ($socials as $bm){
-        $bms[] = array('icon'=>$bm->getVar('icon'),'alt'=>$bm->getVar('alt'),'link'=>str_replace(array('{URL}','{TITLE}','{DESC}'), array($post->permalink(),$post->getVar('title'),TextCleaner::getInstance()->truncate($text, 50)),$bm->getVar('url')));
+    if($xoopsModuleConfig['showbookmarks']){
+        $bms = array();
+        foreach ($socials as $bm){
+            $bms[] = array('icon'=>$bm->getVar('icon'),'alt'=>$bm->getVar('alt'),'link'=>str_replace(array('{URL}','{TITLE}','{DESC}'), array($post->permalink(),$post->getVar('title'),TextCleaner::getInstance()->truncate($text, 50)),$bm->getVar('url')));
+        }
     }
     
     $xoopsTpl->append('posts', array(
